Extract ShoppingListPageContent component because there was a lot of conditional logic in the ShoppingListPage component that was getting out of hand
